requirements.txt
# Backend - API FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Base de données
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Frontend - Flet
flet==0.24.1

# Client HTTP
requests==2.32.3
httpx==0.27.2

# Configuration
python-dotenv==1.0.1

# Utilitaires
python-dateutil==2.9.0

# Graphiques (optionnel)
plotly==5.24.1

# Développement et tests (optionnel)
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0
